[Observability] Drop @kbn/cases-plugin ref from observability-shared
observability_shared referenced the entire @kbn/cases-plugin TS project
(~215 transitive projects) for a single value import: CasesDeepLinkId,
used in one nav switch statement.

Relocate CasesDeepLinkId / ICasesDeepLinkId (a 4-string constant) into
the lightweight @kbn/cases-components package; @kbn/cases-plugin re-exports
them so its public API is unchanged. observability_shared now imports the
id from @kbn/cases-components and drops the @kbn/cases-plugin TS reference
and the `cases` requiredBundle.

Cuts observability-shared's transitive closure ~802 -> ~588 projects.
No runtime behavior change.
