+++ src/platform/packages/shared/kbn-cases-components/index.ts
@@ -14,3 +14,4 @@ export { Tooltip } from './src/tooltip/tooltip';
 export type { CaseTooltipProps, CaseTooltipContentProps } from './src/tooltip/types';
 export { UserActionTitleLink, type UserActionTitleLinkProps } from './src/user_action_title_link';
 export { UserActionTitle, type UserActionTitleProps } from './src/user_action_title';
+export { CasesDeepLinkId, type ICasesDeepLinkId } from './src/deep_link_ids';
