+++ x-pack/solutions/observability/plugins/observability_shared/kibana.jsonc
@@ -23,7 +23,6 @@
       "agentBuilder"
     ],
     "requiredBundles": [
-      "cases",
       "data",
       "inspector",
       "kibanaReact",
