+++ x-pack/solutions/observability/plugins/observability_shared/moon.yml
@@ -19,7 +19,7 @@ project:
 dependsOn:
   - '@kbn/core'
   - '@kbn/kibana-react-plugin'
-  - '@kbn/cases-plugin'
+  - '@kbn/cases-components'
   - '@kbn/i18n'
   - '@kbn/shared-ux-page-kibana-template'
   - '@kbn/i18n-react'
