+++ x-pack/solutions/observability/plugins/observability_shared/tsconfig.json
@@ -15,7 +15,7 @@
   "kbn_references": [
     "@kbn/core",
     "@kbn/kibana-react-plugin",
-    "@kbn/cases-plugin",
+    "@kbn/cases-components",
     "@kbn/i18n",
     "@kbn/shared-ux-page-kibana-template",
     "@kbn/i18n-react",
